requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==24.1.0
aiosqlite==0.20.0
google-genai==1.1.0
notion-client==2.2.1
jinja2==3.1.4
langchain-core>=0.3.0
langchain-google-genai>=2.1.0
numpy>=2.0.0
python-multipart==0.0.9
sqlmodel>=0.0.22
PyJWT>=2.9.0
passlib[bcrypt]>=1.7.4
alembic>=1.13.3
loguru>=0.7.2
